+++ app/javascript/controllers/maps/maplibre_controller.js
@@ -82,6 +82,7 @@ export default class extends Controller {
     // Area selection
     "selectAreaButton",
     "selectionActions",
+    "deleteButtonText",
     "selectedVisitsContainer",
     "selectedVisitsBulkActions",
     // Info display
@@ -1193,6 +1194,9 @@ export default class extends Controller {
   }
   cancelAreaSelection() {
     return this.areaSelectionManager.cancelAreaSelection()
+  }
+  deleteSelectedPoints() {
+    return this.areaSelectionManager.deleteSelectedPoints()
   }
 
   // Visits Manager methods